+++ app/workers/textConverter.js
@@ -4,6 +4,7 @@ import {
     ANSI_BACKGROUND_COLOR_CODES,
     ANSI_FOREGROUND_COLOR_CODES,
     ANSI_CODE_PAGE_437_TABLES,
+    SNDINFO,
 } from '../lib/constants';
 
 const interpolateBackgroundAnsiCode = int => (
@@ -14,25 +15,25 @@ const interpolateForegroundAnsiCode = int => (
     int === Number.isNaN ? null : ANSI_FOREGROUND_COLOR_CODES[int]
 );
 
-const parseAnsiScreen = (data) => {
+const parseAnsiScreen = (input) => {
     const blocks = [];
     let line = [];
 
-    for (let i = 0; i <= data.byteLength; i += 2) {
+    for (let i = 0; i <= input.byteLength; i += 2) {
         if (i !== 0 && i % 160 === 0) {
             blocks.push(line);
             line = [];
         }
 
         // last byte is unimportant but necessary to push the last line
-        if (i === data.byteLength) {
+        if (i === input.byteLength) {
             break;
         }
 
-        const characterByte = data.getUint8(i);
+        const characterByte = input.getUint8(i);
         const character = ANSI_CODE_PAGE_437_TABLES[characterByte];
 
-        const colorByte = data.getUint8(i + 1);
+        const colorByte = input.getUint8(i + 1);
 
         // missings bits are replaced by zeroes
         const colorBits = (`000000000${colorByte.toString(2)}`).substr(-8);
@@ -55,59 +56,65 @@ const parseAnsiScreen = (data) => {
     return blocks;
 };
 
-const decodeText = data => decodeURI(new TextDecoder('utf-8').decode(data).replace(/\u0000/g, ' '));
+const firstDecodeTextMethod = input => decodeURI(new TextDecoder('utf-8').decode(input).replace(/\u0000/g, ' '));
 
 const processNewLines = text => text.split('\n');
 
 onmessage = (message) => {
-    const { wadId, lumpId, data } = message.data;
+    const { wadId, lumpId, input } = message.data;
 
-    console.log(`Converting '${lumpId}' to text (WAD: '${wadId}') ...`);
+    // console.log(`Converting '${lumpId}' to text (WAD: '${wadId}') ...`);
 
-    if (ANSI_LUMPS.includes(lumpId)) {
-        const screenData = parseAnsiScreen(data);
+    let output = null;
+    let convertedFormat = 'text';
 
+    if (lumpId === SNDINFO) {
         postMessage({
             wadId,
             lumpId,
-            text: screenData,
+            ignored: true,
         });
 
         return;
     }
 
-    // rules out certain lumps that are definitively not meant to be rendered as text
-    let decodedText = null;
-    try {
-        decodedText = decodeText(data);
-    } catch (err) {
-        const errorMessage = `Could not convert '${lumpId}' to text (WAD: '${wadId}').`;
-        const id = INVALID_TEXT;
-        const error = {
-            id,
-            message: errorMessage,
-        };
+    if (ANSI_LUMPS.includes(lumpId)) {
+        // ANSI
+        output = parseAnsiScreen(input);
+        convertedFormat = 'ANSI';
+    } else {
+        // text
+        // rules out certain lumps that are definitively not meant to be rendered as text
+        let splitText = null;
+        try {
+            const decodedText = firstDecodeTextMethod(input);
+            splitText = processNewLines(decodedText);
+        } catch (err) {
+            const errorMessage = `Could not convert '${lumpId}' to text (WAD: '${wadId}').`;
+            const id = INVALID_TEXT;
+            const error = {
+                id,
+                message: errorMessage,
+            };
 
-        console.error(error, { err });
+            console.error(error, { err });
 
-        postMessage({
-            wadId,
-            lumpId,
-            text: null,
-            error,
-        });
+            postMessage({
+                wadId,
+                lumpId,
+                error,
+            });
+
+            return;
+        }
+
+        output = splitText;
     }
-
-    const splitText = processNewLines(decodedText);
-    const text = splitText;
-
-
-    console.log({ text });
-
 
     postMessage({
         wadId,
         lumpId,
-        text,
+        output,
+        convertedFormat,
     });
 };